fix(cost): address review — zhipuai alias, split index migrations, org-scoped CTE
- Pricing: alias zai/z-ai -> zhipuai (genai-prices hosts GLM under zhipuai);
  GLM runs were silently unpriced. Fix the alias test to assert real catalog
  pricing instead of self-consistency of two unpriced results.
- Migration: split the 3 CREATE INDEX CONCURRENTLY into one-per-file
  transaction:false migrations (multi-statement caused Postgres 25001), and
  make watcher_id bigint (squawk prefer-bigint-over-int).
- query_sql: add a run_usage org-scoping CTE branch in buildScopedQuery — it
  was falling through to the entity-slug path, leaving the admin-gated ledger
  unqueryable. Add a regression test.
- Ledger: run the usage write on the snapshot 409 path too, so a crash between
  snapshot commit and the ledger write self-heals on a later retry.
- Remove the dead price-override path (lands with the override table in a later
  PR) + soften its docstring; fix a stray NUL byte in the model-aggregation
  key; soften the migration/backfill comments that over-promised; register the
  backfill script as a knip entry.

+++ db/migrations/20260620120000_run_usage.sql
@@ -8,8 +8,9 @@
 -- Why a dedicated table (not events, not runs): the append-only events store
 -- reads through current_event_records (an embedding join we don't want on a
 -- cost hot path), and the runs queue is pruned on retention so it can't be
--- history. Dimensions are denormalized so every rollup is a plain GROUP BY; a
--- sibling events row is emitted separately for the metric_series sparklines.
+-- history. Dimensions are denormalized so every rollup is a plain GROUP BY. (A
+-- sibling append-only events row for the metric_series sparklines is a planned
+-- follow-up with the showback UI; it is NOT wired in this PR.)
 --
 -- usd is NULL + unpriced=true when no model in the price catalog (and no org
 -- override) can price the run — never a fake $0. The token buckets are always
@@ -22,7 +23,7 @@ CREATE TABLE IF NOT EXISTS public.run_usage (
     conversation_id text NOT NULL,
     run_id bigint NOT NULL,
     user_id text,
-    watcher_id integer,
+    watcher_id bigint,
     source text,
     provider text,
     model text,
